% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.R
\name{cardBackup}
\alias{cardBackup}
\title{Card Backup Card}
\usage{
cardBackup(card_id, filename = "temp.json", read_file = TRUE)
}
\arguments{
\item{card_id}{card id}

\item{filename}{export filename}

\item{read_file}{either \code{TRUE} or \code{FALSE}}
}
\value{

}
\description{
backup-card: export a specified card as a \code{json} blob to a file or stdout that can be used to restore it later
}
