% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.R
\name{cardExport}
\alias{cardExport}
\title{Card Export Data}
\usage{
cardExport(card_id, filename = "temp.csv", read_file = TRUE)
}
\arguments{
\item{card_id}{card id}

\item{filename}{export filename}

\item{read_file}{either \code{TRUE} or \code{FALSE}}
}
\value{

}
\description{
download-card-data: returns card data as a \code{csv} file
}
