% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.R
\name{dfList}
\alias{dfList}
\title{Dataflow List}
\usage{
dfList(
  df_id = NULL,
  filename = NULL,
  limit = 1,
  name_like = NULL,
  offset = NULL,
  time = NULL,
  versions = NULL
)
}
\arguments{
\item{df_id}{dataflow id}

\item{filename}{export filename}

\item{limit}{the number of dataflows/executions to return}

\item{name_like}{name filter}

\item{offset}{the offset into the pagination}

\item{time}{the last start time in millis}

\item{versions}{list dataflow versions}
}
\value{

}
\description{
list-dataflow: lists domo dataflows
}
