% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.R
\name{dfSetProperties}
\alias{dfSetProperties}
\title{Dataflow Set Properties}
\usage{
dfSetProperties(
  definition,
  df_id,
  df_name = NULL,
  enable = NULL,
  create = NULL,
  subscribe = NULL
)
}
\arguments{
\item{definition}{definition}

\item{df_id}{dataflow id}

\item{df_name}{name}

\item{enable}{enable (true/false)}

\item{create}{create}

\item{subscribe}{subscriber (true/false)}
}
\value{

}
\description{
set-dataflow-properties: set a domo dataflow property
}
