% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domoclir-package.R
\docType{package}
\name{domoclir-package}
\alias{domoclir}
\alias{domoclir-package}
\title{domoclir: An R wrapper for Domo's Java CLI}
\description{
An R wrapper for Domo's Java CLI.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimmyg3g/domoclir}
  \item Report bugs at \url{https://github.com/jimmyg3g/domoclir/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Glenn \email{jimmykglenn@gmail.com}

}
\keyword{internal}
