% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.R
\name{dsCreate}
\alias{dsCreate}
\title{Dataset Create}
\usage{
dsCreate(
  description = NULL,
  ds_id = NULL,
  name,
  schema_file,
  type,
  index = NULL
)
}
\arguments{
\item{description}{Optional, dataset description (optional)}

\item{ds_id}{Optional, dataset id (optional) - if provided, schema of dataset is updated}

\item{name}{dataset name}

\item{schema_file}{filename for schema definition}

\item{type}{dataset type}

\item{index}{Optional, index dataset after schema change (optional)}
}
\value{

}
\description{
create-dataset: Creates a Domo Dataset
Many variations to how this works depending on supplied params, need more documentation!
}
