% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.R
\name{dsDeriveSchema}
\alias{dsDeriveSchema}
\title{Dataset Derive Schema}
\usage{
dsDeriveSchema(data, row_limit = 1000, schema_file = "schemafile.json")
}
\arguments{
\item{data}{either a \code{csv} file or \code{data.frame}}

\item{row_limit}{Optional, rows to parse for schema detection, default = 1000}

\item{schema_file}{filename for schema definition}
}
\value{

}
\description{
derive-schema: derives a dataset schema from an input file
}
