% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.R
\name{dsExport}
\alias{dsExport}
\title{Dataset Export}
\usage{
dsExport(
  ds_id,
  filename = NULL,
  read_file = TRUE,
  query = NULL,
  queryfile = NULL
)
}
\arguments{
\item{ds_id}{dataset id}

\item{filename}{export filename}

\item{read_file}{either \code{TRUE} or \code{FALSE}}

\item{query}{optional query}

\item{queryfile}{query filename}
}
\value{

}
\description{
export-data: exports datset to a \code{csv} file
}
