% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.R
\name{dsExportVersion}
\alias{dsExportVersion}
\title{Dataset Export Version}
\usage{
dsExportVersion(ds_id, version, filename = "temp.csv", read_file = TRUE)
}
\arguments{
\item{ds_id}{dataset id}

\item{version}{version}

\item{filename}{export filename}

\item{read_file}{either \code{TRUE} or \code{FALSE}}
}
\value{

}
\description{
export-version: exports a specific dataset version to a \code{csv} file
}
