% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.R
\name{dsGetDetails}
\alias{dsGetDetails}
\title{Dataset Get Details}
\usage{
dsGetDetails(ds_id, filename = NULL, read_file = TRUE)
}
\arguments{
\item{ds_id}{dataset id}

\item{filename}{}
}
\value{

}
\description{
get-detaset: returns dataset details for the specified id
}
