% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.R
\name{dsGetVersion}
\alias{dsGetVersion}
\title{Dataset Get Version}
\usage{
dsGetVersion(ds_id, format = "wide")
}
\arguments{
\item{ds_id}{dataset id}

\item{format}{either \code{wide} or \code{long}}
}
\value{

}
\description{
get-version: returns the specified Domo dataset version
}
