% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.R
\name{dsMoveData}
\alias{dsMoveData}
\title{Move Data}
\usage{
dsMoveData(source, destination, append = NULL)
}
\arguments{
\item{source}{source dataset}

\item{destination}{destination dataset}

\item{append}{c(TRUE, NULL) append to existing data}
}
\value{

}
\description{
move-data: Moves dataa from one Domo dataset to another
}
