% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.R
\name{dsSetSchema}
\alias{dsSetSchema}
\title{Dataset Set Schema}
\usage{
dsSetSchema(ds_id, schema_file = "schemafile.json")
}
\arguments{
\item{ds_id}{dataset id}

\item{schema_file}{schema definition file}
}
\value{

}
\description{
set-schema: sets the specified schema defintion on a dataset. This command does not work with data views (use \code{create-dataview} for data views)
}
